requirements/base.txt
django==4.2.11
psycopg2-binary==2.9.9
pillow==10.2.0
redis==5.0.2
python-environ==0.4.54
django_redis==5.4.0
django-jazzmin==2.6.0
django-filter==23.5
django-modeltranslation==0.18.11
djangorestframework==3.14.0
django-cors-headers==4.3.1
drf-yasg==1.21.7
django-mptt==0.16.0
